app/requirements_dev.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
astroid==2.15.1 ; python_full_version >= '3.7.2'
attrs==22.2.0 ; python_version >= '3.6'
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7 ; python_version >= '3.6'
cfgv==3.3.1 ; python_full_version >= '3.6.1'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
coverage==7.2.2
dill==0.3.6 ; python_version >= '3.11'
distlib==0.3.6
dparse==0.6.2 ; python_version >= '3.5'
filelock==3.10.7 ; python_version >= '3.7'
flake8==6.0.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
identify==2.5.22 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
isort==5.12.0 ; python_full_version >= '3.8.0'
lazy-object-proxy==1.9.0 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
mypy==1.1.1
mypy-extensions==1.0.0 ; python_version >= '3.5'
nodeenv==1.7.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.0 ; python_version >= '3.7'
pathspec==0.11.1 ; python_version >= '3.7'
platformdirs==3.2.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pre-commit==3.2.1
pycodestyle==2.10.0 ; python_version >= '3.6'
pyflakes==3.0.1 ; python_version >= '3.6'
pylint==2.17.1
pytest==7.2.2
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rfc3986[idna2008]==1.5.0
ruamel.yaml==0.17.21 ; python_version >= '3'
safety==2.3.4
setuptools==67.6.1 ; python_version >= '3.7'
sniffio==1.3.0 ; python_version >= '3.7'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomlkit==0.11.7 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
virtualenv==20.21.0 ; python_version >= '3.7'
wrapt==1.15.0 ; python_version >= '3.11'
fastapi==0.95.0
pydantic==1.10.7 ; python_version >= '3.7'
python-dotenv==0.21.1
starlette==0.26.1 ; python_version >= '3.7'
uvicorn==0.21.1
